requirements.txt
hyperdiv
requests
Authlib
python-dotenv